seperate camera scan controller into two classes
